Claim_Bot/requirements.txt
streamlit
python-dotenv
openai
pytesseract
pdfplumber
pandas
